update python exposure
