+++ kratos/python/add_container_expression_to_python.cpp
@@ -16,15 +16,14 @@
 // External includes
 #include <pybind11/stl.h>
 #include <pybind11/operators.h>
-#include <pybind11/numpy.h>
 
 // Project includes
-#include "includes/define_python.h"
-#include "expression/container_data_io.h"
-#include "expression/container_expression.h"
-#include "expression/specialized_container_expression.h"
-#include "add_container_expression_to_python_utils.h"
 #include "add_expression_io_to_python.h"
+#include "expression/arithmetic_operators.h"
+#include "expression/c_array_expression_io.h"
+#include "expression/container_expression.h"
+#include "includes/define_python.h"
+#include "numpy_utils.h"
 
 // Include base h
 #include "add_container_expression_to_python.h"
@@ -32,6 +31,102 @@
 namespace Kratos::Python
 {
 
+template<class TContainerType>
+void AddContainerExpressionToPython(pybind11::module& m, const std::string& rName)
+{
+    namespace py = pybind11;
+
+    using container_expression_holder_base = ContainerExpression<TContainerType>;
+    py::class_<container_expression_holder_base, typename container_expression_holder_base::Pointer>(m, rName.c_str())
+        .def(py::init<ModelPart&>(), py::arg("model_part"))
+        .def("CopyFrom", &container_expression_holder_base::CopyFrom, py::arg("origin_container_expression"))
+        .def("MoveFrom", [](container_expression_holder_base& rSelf, py::array_t<int>& rData){
+            KRATOS_ERROR_IF(rData.ndim() == 0) << "Passed data is not compatible.\n";
+
+            // dimension of the numpy array is always one dimension greater than the kratos stored dimension for each
+            // entity. That is because, first dimension of the numpy array shows how many entities are there
+            // in the numpy array to be read in. If the numpy array dimension is [45, 3, 4] then it shows
+            // there are 45 entities each having matrices of shape [3, 4].
+            std::vector<int> shape(rData.ndim() - 1);
+            std::copy(rData.shape() + 1, rData.shape() + rData.ndim(), shape.begin());
+
+            rSelf.MoveFrom(rData.mutable_data(),
+                           rData.shape()[0],
+                           shape.data(),
+                           shape.size());
+        }, py::arg("numpy_int_array").noconvert())
+        .def("MoveFrom", [](container_expression_holder_base& rSelf, py::array_t<double>& rData){
+            KRATOS_ERROR_IF(rData.ndim() == 0) << "Passed data is not compatible.\n";
+
+            // dimension of the numpy array is always one dimension greater than the kratos stored dimension for each
+            // entity. That is because, first dimension of the numpy array shows how many entities are there
+            // in the numpy array to be read in. If the numpy array dimension is [45, 3, 4] then it shows
+            // there are 45 entities each having matrices of shape [3, 4].
+            std::vector<int> shape(rData.ndim() - 1);
+            std::copy(rData.shape() + 1, rData.shape() + rData.ndim(), shape.begin());
+
+            rSelf.MoveFrom(rData.mutable_data(),
+                           rData.shape()[0],
+                           shape.data(),
+                           shape.size());
+        }, py::arg("numpy_double_array").noconvert())
+        .def("SetExpression", &container_expression_holder_base::SetExpression)
+        .def("HasExpression", &container_expression_holder_base::HasExpression)
+        .def("GetExpression", &container_expression_holder_base::pGetExpression)
+        .def("GetModelPart", py::overload_cast<>(&container_expression_holder_base::GetModelPart), py::return_value_policy::reference)
+        .def("GetContainer", py::overload_cast<>(&container_expression_holder_base::GetContainer), py::return_value_policy::reference)
+        .def("GetItemShape", &container_expression_holder_base::GetItemShape)
+        .def("GetItemComponentCount", &container_expression_holder_base::GetItemComponentCount)
+        .def("Slice",
+             &container_expression_holder_base::Slice,
+             py::arg("offset"),
+             py::arg("stride"))
+        .def("Reshape",
+             &container_expression_holder_base::Reshape,
+             py::arg("new_shape"))
+        .def("Comb",
+             [](container_expression_holder_base& rSelf,
+                const container_expression_holder_base& rOther)
+                {return rSelf.Comb(rOther);},
+             py::arg("other"))
+        .def("Comb",
+             [](container_expression_holder_base& rSelf,
+                const std::vector<typename container_expression_holder_base::Pointer>& rOthers)
+                {return rSelf.Comb(rOthers);},
+             py::arg("others"))
+        .def("Evaluate", [](const container_expression_holder_base& rSelf){
+            const auto& r_shape = rSelf.GetItemShape();
+            auto array = AllocateNumpyArray<double>(rSelf.GetContainer().size(), r_shape);
+            CArrayExpressionIO::Write(rSelf, array.mutable_data(), array.size());
+            return array;
+        })
+        .def("Clone", &container_expression_holder_base::Clone)
+        .def("Scale", [](const container_expression_holder_base& rSelf, const container_expression_holder_base& rOther){auto copy = rSelf; copy.SetExpression(Scale(rSelf.pGetExpression(), rOther.pGetExpression())); return copy;})
+        .def("__add__", [](const container_expression_holder_base& rSelf, const container_expression_holder_base& rOther) { return rSelf + rOther; })
+        .def("__iadd__", [](container_expression_holder_base& rSelf, const container_expression_holder_base& rOther) { rSelf = rSelf + rOther; return rSelf; })
+        .def("__add__", [](const container_expression_holder_base& rSelf, const double Value) { return rSelf + Value; })
+        .def("__iadd__", [](container_expression_holder_base& rSelf, const double Value) { rSelf = rSelf + Value; return rSelf; })
+        .def("__sub__", [](const container_expression_holder_base& rSelf, const container_expression_holder_base& rOther) { return rSelf - rOther; })
+        .def("__isub__", [](container_expression_holder_base& rSelf, const container_expression_holder_base& rOther) { rSelf = rSelf - rOther; return rSelf; })
+        .def("__sub__", [](const container_expression_holder_base& rSelf, const double Value) { return rSelf - Value; })
+        .def("__isub__", [](container_expression_holder_base& rSelf, const double Value) { rSelf = rSelf - Value; return rSelf; })
+        .def("__mul__", [](const container_expression_holder_base& rSelf, const container_expression_holder_base& rOther) { return rSelf * rOther; })
+        .def("__imul__", [](container_expression_holder_base& rSelf, const container_expression_holder_base& rOther) { rSelf = rSelf * rOther; return rSelf; })
+        .def("__mul__", [](const container_expression_holder_base& rSelf, const double Value) { return rSelf * Value; })
+        .def("__imul__", [](container_expression_holder_base& rSelf, const double Value) { rSelf = rSelf * Value; return rSelf; })
+        .def("__truediv__", [](const container_expression_holder_base& rSelf, const container_expression_holder_base& rOther) { return rSelf / rOther; })
+        .def("__itruediv__", [](container_expression_holder_base& rSelf, const container_expression_holder_base& rOther) { rSelf = rSelf / rOther; return rSelf; })
+        .def("__truediv__", [](const container_expression_holder_base& rSelf, const double Value) { return rSelf / Value; })
+        .def("__itruediv__", [](container_expression_holder_base& rSelf, const double Value) { rSelf = rSelf / Value; return rSelf; })
+        .def("__pow__", [](container_expression_holder_base& rSelf, const container_expression_holder_base& rInput) { container_expression_holder_base result(rSelf.GetModelPart()); result = Power(rSelf, rInput); return result; })
+        .def("__ipow__", [](container_expression_holder_base& rSelf, const container_expression_holder_base& rInput) { rSelf = Power(rSelf, rInput); return rSelf; })
+        .def("__pow__", [](container_expression_holder_base& rSelf, const double Value) { container_expression_holder_base result(rSelf.GetModelPart()); result = Power(rSelf, Value); return result; })
+        .def("__ipow__", [](container_expression_holder_base& rSelf, const double Value) { rSelf = Power(rSelf, Value); return rSelf; })
+        .def("__neg__", [](container_expression_holder_base& rSelf) { return rSelf *= -1.0; })
+        .def("PrintData", &container_expression_holder_base::PrintData)
+        .def("__str__", &container_expression_holder_base::Info)
+        ;
+}
 
 void  AddContainerExpressionToPython(pybind11::module& m)
 {
@@ -40,11 +135,6 @@ void  AddContainerExpressionToPython(pybind11::module& m)
     AddContainerExpressionToPython<ModelPart::NodesContainerType>(container_exp_sub_module, "NodalExpression");
     AddContainerExpressionToPython<ModelPart::ConditionsContainerType>(container_exp_sub_module, "ConditionExpression");
     AddContainerExpressionToPython<ModelPart::ElementsContainerType>(container_exp_sub_module, "ElementExpression");
-
-    AddSpecializedContainerExpressionToPython<ModelPart::NodesContainerType, ContainerDataIOTags::Historical>(container_exp_sub_module, "HistoricalExpression");
-    AddSpecializedContainerExpressionToPython<ModelPart::NodesContainerType, ContainerDataIOTags::NonHistorical>(container_exp_sub_module, "NodalNonHistoricalExpression");
-    AddSpecializedContainerExpressionToPython<ModelPart::ConditionsContainerType, ContainerDataIOTags::NonHistorical>(container_exp_sub_module, "ConditionNonHistoricalExpression");
-    AddSpecializedContainerExpressionToPython<ModelPart::ElementsContainerType, ContainerDataIOTags::NonHistorical>(container_exp_sub_module, "ElementNonHistoricalExpression");
 
     AddExpressionIOToPython(container_exp_sub_module);
 }
